+++ cmd/sunani/canvas.go
@@ -11,8 +11,8 @@ type Canvas struct {
 
 	window *glfw.Window
 
-	r, g, b, a uint32
-	points     []uint32
+	r, g, b, a int
+	points     []int
 }
 
 func NewCanvas() *Canvas {
@@ -68,7 +68,7 @@ func (c *Canvas) Begin() {
 	gl.Viewport(0, 0, int32(fw), int32(fh))
 }
 
-func (c *Canvas) Color(r, g, b, a uint32) {
+func (c *Canvas) Color(r, g, b, a int) {
 	if !c.IsEnabled() {
 		errlog("sunani canvas.color was called, but Canvas API is not enabled.\nExport snunani_canvas_init to enable this API.")
 		return
@@ -81,7 +81,7 @@ func (c *Canvas) Color(r, g, b, a uint32) {
 	gl.Color4ub(uint8(r), uint8(g), uint8(b), uint8(a))
 }
 
-func (c *Canvas) Line(x1, y1 uint32, x2, y2 uint32) {
+func (c *Canvas) Line(x1, y1 int, x2, y2 int) {
 	if !c.IsEnabled() {
 		errlog("sunani canvas.line was called, but Canvas API is not enabled.\nExport snunani_canvas_init to enable this API.")
 		return
@@ -97,7 +97,7 @@ func (c *Canvas) Line(x1, y1 uint32, x2, y2 uint32) {
 	gl.End()
 }
 
-func (c *Canvas) Rect(x, y uint32, w, h uint32) {
+func (c *Canvas) Rect(x, y int, w, h int) {
 	if !c.IsEnabled() {
 		errlog("sunani canvas.rect was called, but Canvas API is not enabled.\nExport snunani_canvas_init to enable this API.")
 		return
@@ -115,7 +115,7 @@ func (c *Canvas) Rect(x, y uint32, w, h uint32) {
 	gl.End()
 }
 
-func (c *Canvas) FillRect(x, y uint32, w, h uint32) {
+func (c *Canvas) FillRect(x, y int, w, h int) {
 	if !c.IsEnabled() {
 		errlog("sunani canvas.fill_rect was called, but Canvas API is not enabled.\nExport snunani_canvas_init to enable this API.")
 		return
@@ -133,7 +133,7 @@ func (c *Canvas) FillRect(x, y uint32, w, h uint32) {
 	gl.End()
 }
 
-func (c *Canvas) polygon(points []uint32) {
+func (c *Canvas) polygon(points []int) {
 	n := len(points) / 2
 	if n < 2 {
 		return
@@ -153,12 +153,12 @@ func (c *Canvas) polygon(points []uint32) {
 	)
 }
 
-func cross(ax, ay, bx, by, cx, cy uint32) int32 {
-	return int32((bx-ax)*(cy-ay) - (by-ay)*(cx-ax))
+func cross(ax, ay, bx, by, cx, cy int) int {
+	return (bx-ax)*(cy-ay) - (by-ay)*(cx-ax)
 }
 
-func pointInTriangle(px, py uint32,
-	ax, ay, bx, by, cx, cy uint32) bool {
+func pointInTriangle(px, py int,
+	ax, ay, bx, by, cx, cy int) bool {
 
 	c1 := cross(px, py, ax, ay, bx, by)
 	c2 := cross(px, py, bx, by, cx, cy)
@@ -168,9 +168,9 @@ func pointInTriangle(px, py uint32,
 		(c1 <= 0 && c2 <= 0 && c3 <= 0)
 }
 
-func triangulatePolygon(points []uint32) []uint32 {
-	tris := []uint32{}
-	verts := append([]uint32{}, points...)
+func triangulatePolygon(points []int) []int {
+	tris := []int{}
+	verts := append([]int{}, points...)
 
 	for len(verts) >= 6 {
 		n := len(verts) / 2
@@ -222,7 +222,7 @@ func triangulatePolygon(points []uint32) []uint32 {
 	return tris
 }
 
-func (c *Canvas) triangle(x1, y1, x2, y2, x3, y3 uint32) {
+func (c *Canvas) triangle(x1, y1, x2, y2, x3, y3 int) {
 	gl.Color4ub(uint8(c.r), uint8(c.g), uint8(c.b), uint8(c.a))
 	gl.Begin(gl.TRIANGLES)
 	gl.Vertex2i(int32(x1), int32(y1))
@@ -231,7 +231,7 @@ func (c *Canvas) triangle(x1, y1, x2, y2, x3, y3 uint32) {
 	gl.End()
 }
 
-func (c *Canvas) fillPolygon(points []uint32) {
+func (c *Canvas) fillPolygon(points []int) {
 	tris := triangulatePolygon(points)
 
 	for i := 0; i < len(tris); i += 6 {
@@ -252,10 +252,10 @@ func (c *Canvas) Path() {
 		panic("window is nil")
 	}
 
-	c.points = []uint32{}
+	c.points = []int{}
 }
 
-func (c *Canvas) Vertex(x, y uint32) {
+func (c *Canvas) Vertex(x, y int) {
 	if !c.IsEnabled() {
 		errlog("sunani canvas.vertex was called, but Canvas API is not enabled.\nExport snunani_canvas_init to enable this API.")
 		return
